회원가입 기능 구현 및 JWT 토큰 로직 구현중(검증필요)

+++ backend/src/main/java/tazoria/banimo/user/repository/UserRepository.java
@@ -5,8 +5,7 @@ import tazoria.banimo.user.entity.UserEntity;
 
 import java.util.Optional;
 
-
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByUsername(String username);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
